Improve the naming of Bundle ID lookup api
It also now throws rather than use fatalError
