+++ Sources/AppStoreConnectCLI/Commands/BundleIds/ModifyBundleIdCommand.swift
@@ -25,7 +25,7 @@ struct ModifyBundleIdCommand: CommonParsableCommand {
         let api = try makeClient()
 
         _ = try api
-            .internalId(matching: identifier)
+            .bundleIdResourceId(matching: identifier)
             .flatMap { internalId in
                 api.request(APIEndpoint.modifyBundleId(id: internalId, name: self.name)).eraseToAnyPublisher()
             }
