+++ Sources/AppStoreConnectCLI/Commands/BundleIds/ReadBundleIdCommand.swift
@@ -21,7 +21,7 @@ struct ReadBundleIdCommand: CommonParsableCommand {
         let api = try makeClient()
 
         _ = try api
-            .internalId(matching: identifier)
+            .bundleIdResourceId(matching: identifier)
             .flatMap { internalId in
                 api.request(APIEndpoint.readBundleIdInformation(id: internalId)).eraseToAnyPublisher()
             }
